requirements.txt
anyio==3.7.1
asgiref~=3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel~=2.12.1
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Django~=4.2.4
django-allauth~=0.55.2
django-crontab==0.7.1
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-redis==5.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-theme-material-kit==1.0.19
djangorestframework~=3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
exceptiongroup==1.1.2
fonttools==4.42.1
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.22.0
httpx~=0.24.1
idna==3.4
inflection==0.5.1
injector==0.21.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kubernetes~=26.1.0
MarkupSafe==2.1.3
marshmallow~=3.20.1
mysqlclient==2.0.3
numpy==1.25.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
panda==0.3.1
pandas~=2.1.0
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydyf==0.7.0
PyJWT==2.8.0
pyminizip==0.2.6
pyparsing==3.1.1
PyPDF2~=3.0.1
pyphen==0.14.0
python-dateutil~=2.8.2
python-dotenv~=1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests~=2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.0
rsa==4.9
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
weasyprint~=59.0
webencodings==0.5.1
websocket-client==1.6.1
whitenoise==6.5.0
zopfli==0.2.2
